pyproject.toml
[project]
name = "chuk-mcp-celestial"
version = "0.2"
description = "MCP server for US Navy astronomical and celestial data - moon phases, eclipses, sun/moon rise/set times"
authors = [
    {name = "Chris Hay", email = "chris@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chuk-mcp-server>=0.8",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "chuk-virtual-fs>=0.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "types-PyYAML>=6.0.0",
]
skyfield = [
    "skyfield>=1.48",
    "pyyaml>=6.0",
    "boto3>=1.28.0",
    "aioboto3>=12.0.0",
    "python-dotenv>=1.0.0",
]
all = [
    "chuk-mcp-celestial[dev,skyfield]",
]

[project.scripts]
chuk-mcp-celestial = "chuk_mcp_celestial.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test",
    "slow: mark test as slow running",
    "network: mark test as requiring network access",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
disable_error_code = ["arg-type"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
